型マッピングとテスト機能の強化
ParquetWriterManager.csにおいて、TypeMappingManagerを使用してフィールドの型を動的に取得するロジックを追加し、データ変換メソッドを強化しました。Program.csでは、コマンドライン引数に基づくテストモードの実行ロジックを追加しました。TypeMappingManager.csの初期化処理を改善し、新しいレコードタイプマッピングを追加しました。また、各レコードタイプマッピングクラスを新たに追加し、データ型のマッピングを強化しました。TypeMappingProgress.mdを更新し、進捗状況を反映しました。

+++ JVParquet/TypeMapping/RecordTypes/BNRecordTypeMapping.cs
@@ -35,8 +35,8 @@ namespace JVParquet.TypeMapping.RecordTypes
             // 本年・累計成績情報（配列展開される）
             // HonRuikei[0] - 本年成績
             { "HonRuikei_0__SetYear", typeof(int) },              // 設定年
-            { "HonRuikei_0__HonSyokinTotal", typeof(decimal) },   // 本賞金合計
-            { "HonRuikei_0__FukaSyokin", typeof(decimal) },       // 付加賞金合計
+            { "HonRuikei_0__HonSyokinTotal", typeof(int) },   // 本賞金合計
+            { "HonRuikei_0__FukaSyokin", typeof(int) },       // 付加賞金合計
             { "HonRuikei_0__ChakuKaisu_0", typeof(int) },         // 着回数[0] - 1着
             { "HonRuikei_0__ChakuKaisu_1", typeof(int) },         // 着回数[1] - 2着
             { "HonRuikei_0__ChakuKaisu_2", typeof(int) },         // 着回数[2] - 3着
@@ -46,8 +46,8 @@ namespace JVParquet.TypeMapping.RecordTypes
 
             // HonRuikei[1] - 累計成績
             { "HonRuikei_1__SetYear", typeof(int) },              // 設定年
-            { "HonRuikei_1__HonSyokinTotal", typeof(decimal) },   // 本賞金合計
-            { "HonRuikei_1__FukaSyokin", typeof(decimal) },       // 付加賞金合計
+            { "HonRuikei_1__HonSyokinTotal", typeof(int) },   // 本賞金合計
+            { "HonRuikei_1__FukaSyokin", typeof(int) },       // 付加賞金合計
             { "HonRuikei_1__ChakuKaisu_0", typeof(int) },         // 着回数[0] - 1着
             { "HonRuikei_1__ChakuKaisu_1", typeof(int) },         // 着回数[1] - 2着
             { "HonRuikei_1__ChakuKaisu_2", typeof(int) },         // 着回数[2] - 3着
